Move `getTestRunFinalResult` to utils
